Added public key compression specs (#137)
## What has changed?

I've added detailed specs for the implementation of public key compression and decompression. The specifications detail the use of the following `multiformat` features:

- `multibase`
- `multicodec`
- `unsigned-varint`

`multiformat` is used to ensure that the implementation has as much flexibility and robustness as feasible.

## Why make the change?

The usage of key de/compression is outside the typical usage of public keys and requires a degree of background knowledge to correctly implement. The purpose of this specification change is to provide this needed background knowledge.

Please also see

- https://github.com/status-im/status-go/issues/1937
- https://github.com/status-im/status-go/pull/1990

---

* Added Public key compression specs

* Added recommendation for encoding type of compressed keys

* Added unrecognised words to wordlist

* Add multibase to the wordlist

* Added a basic example of the multiformat EC key compression concept

* Added parsable to wordlist

* Hex is the only Lingua Franca we need

* Language to make pk de/compression SHOULD implement

* Added terms to glossary explaining key de/compression

* Change terminology from compress to serialise

* Added rationale for public key compression

* Added deserialization to the wordlist

* Concise sentence

* Added url to the wordlist